Added support for adding new category, subcategory and adjusted how we handle adding new supplier at the moment.
Also introduced some anti-double-postback mechanism (based on generation of current time and checking it on page refresh while sending the data), but it does not counter pop-up for "resending the data". So a different kind of mechanism will have to be used in the future.

Current idea on mechanism: meta-tag refresh with additional parameter to URL - redirection mechanism will should find out current url, strip it from previous "pbstatus=" and set it again (url parameter for some sort of Message class (not written yet)).